Implemented switching between 2d and 2d canvases
