+++ src/animweb/helpers/Constants.ts
@@ -10,4 +10,9 @@ const Constants = {
   Infinity: 5000,
 }
 
+export enum RenderingModes {
+  _2D = '2D',
+  _3D = '3D',
+}
+
 export default Constants
