+++ src/main.ts
@@ -1,5 +1,5 @@
 import NumberPlane from './animweb/AnimObjects/NumberPlane'
-import Scene from './animweb/Scene'
+import Scene2D from './animweb/Scene2D'
 import Point from './animweb/AnimObjects/Point'
 import Line, { Lines } from './animweb/AnimObjects/Line'
 import { Width, Height } from './animweb/helpers/Dimensions'
@@ -12,11 +12,15 @@ import Color from './animweb/helpers/Color'
 import { Transitions } from './animweb/Transition'
 import Text, { TextStyle } from './animweb/AnimObjects/Text'
 import AnimObject, { Observables, AnimObjects } from './animweb/AnimObject'
-import Constants from './animweb/helpers/Constants'
+import Constants, { RenderingModes } from './animweb/helpers/Constants'
 import ImplicitCurve from './animweb/AnimObjects/ImplicitCurve'
 import LaTeX from './animweb/AnimObjects/LaTeX'
 import Matrix from './animweb/helpers/Matrix'
 import Vector, { Vectors } from './animweb/AnimObjects/Vector'
+import ThreeDPlane from './animweb/AnimObjects/3D/3DPlane'
+import Scene3D from './animweb/Scene3D'
+import { EditorView, basicSetup } from 'codemirror'
+import { javascript } from '@codemirror/lang-javascript'
 
 declare global {
   interface Window {
@@ -27,7 +31,21 @@ declare global {
   }
 }
 
-let scene = new Scene(Width.full, Height.full, Colors.gray0)
+let defaultCode = `// code your animation here`
+let editor = new EditorView({
+  //@ts-ignore
+  parent: document.querySelector('.codemirror-editor-container'),
+  doc: defaultCode,
+  extensions: [basicSetup, javascript(), EditorView.lineWrapping],
+})
+
+export type Scene = Scene2D | Scene3D
+let scene: Scene
+let scene2D = new Scene2D(Width.full, Height.full, Colors.gray0, editor)
+let scene3D = new Scene3D(Width.full, Height.full, Colors.gray0, editor)
+scene3D.hide()
+scene2D.show()
+scene = scene2D
 
 let WebAnim = {
   // Basic classes/functions
@@ -39,8 +57,20 @@ let WebAnim = {
   Height,
   wait: async (config: any) => scene.wait(config),
   show: (config: any) => scene.add(config),
+  render: (mode: '3D' | '2D' = '2D') => {
+    if (mode == '3D') {
+      scene2D.hide()
+      scene3D.show()
+      scene = scene3D
+    } else {
+      scene3D.hide()
+      scene2D.show()
+      scene = scene2D
+    }
+  },
   // AnimObjects
   NumberPlane: (config: any) => new NumberPlane({ ...config, scene }),
+  ThreeDPlane: (config: any) => new ThreeDPlane({ ...config, scene }),
   Line: (config: any) => new Line({ ...config, scene }),
   Point: (config: any) => new Point({ ...config, scene }),
   Curve: (config: any) => new Curve({ ...config, scene }),
